renamed all occurances of pybamm-cookiecutter to pybamm-cookie
