+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
       - id: trailing-whitespace
       - id: end-of-file-fixer
       - id: check-yaml
-        files: "^(src/pybamm_cookiecutter|tests$)"
+        files: "^(src/pybamm_cookie|tests$)"
       - id: check-added-large-files
       - id: mixed-line-ending
 
@@ -38,7 +38,7 @@ repos:
     rev: v0.5.7
     hooks:
       - id: ruff
-        files: "^(src/pybamm_cookiecutter|tests$)"
+        files: "^(src/pybamm_cookie|tests$)"
         args: ["--fix", "--show-fixes"]
 
   # Documentation files
@@ -62,7 +62,7 @@ repos:
     rev: v1.11.1
     hooks:
       - id: mypy
-        files: "^(src/pybamm_cookiecutter|tests$)"
+        files: "^(src/pybamm_cookie|tests$)"
 
 exclude:
   CODE-OF-CONDUCT.md
