+++ docs/conf.py
@@ -1,9 +1,9 @@
 import importlib.metadata
 
-project = "pybamm-cookiecutter"
+project = "pybamm-cookie"
 copyright = "2018-2023, The PyBaMM Team"
 author = "Agriya Khetarpal"
-version = release = importlib.metadata.version("pybamm_cookiecutter")
+version = release = importlib.metadata.version("pybamm_cookie")
 
 extensions = [
     "myst_parser",
