+++ src/pybamm_cookie/__main__.py
@@ -1,0 +1,4 @@
+if __name__ == "__main__":
+    from pybamm_cookie.cli import pybamm_cookie_cli
+
+    pybamm_cookie_cli()
