--- src/pybamm_cookiecutter/__main__.py
@@ -1,4 +1,0 @@
-if __name__ == "__main__":
-    from pybamm_cookiecutter.cli import pybamm_cookiecutter_cli
-
-    pybamm_cookiecutter_cli()
