+++ template/{{ project_name }}/src/{{project_slug}}/entry_point.py.jinja
@@ -39,7 +39,7 @@ from typing import Callable
 
 class EntryPoint(Mapping):
     """
-    Dict-like interface for accessing parameter sets and models through entry points in cookiecutter template.
+    Dict-like interface for accessing parameter sets and models through entry points in copier template.
     Access via :py:data:`{{project_slug}}.parameter_sets` for parameter_sets
     Access via :py:data:`{{project_slug}}.Model` for Models
