+++ template/{{ project_name }}/{% if ci == 'github' %}.github{% endif %}/workflows/test_on_push.yml
@@ -33,7 +33,7 @@ jobs:
     name:
       Template Generation/Project Tests (${{ matrix.os }} / Python ${{ matrix.python-version }})
     steps:
-      - name: Checkout pybamm-cookiecutter
+      - name: Checkout pybamm-cookie
         uses: actions/checkout@v4
         with:
           fetch-depth: 0
@@ -74,7 +74,7 @@ jobs:
     name: Doctests (ubuntu-latest / Python 3.12)
 
     steps:
-      - name: Check out pybamm-cookiecutter repository
+      - name: Check out pybamm-cookie repository
         uses: actions/checkout@v4
 
       - name: Set up Python
